[Backend][Updater] Renamed facultyCodes to organisationConstants and moved universityOrgId there. Refactored updater to use universityOrgId from new location organisationConstants
